+++ examples/helcim/helcim.ts
@@ -138,8 +138,8 @@ function _buildRefundRequest(connectorTransactionId: string): types.IPaymentServ
 function _buildRefundGetRequest(): types.IRefundServiceGetRequest {
     return {
         "merchantRefundId": "probe_refund_001",  // Identification.
-        "connectorTransactionId": "12345",
-        "refundId": "probe_refund_id_001"
+        "connectorTransactionId": "probe_connector_txn_001",
+        "refundId": "probe_refund_id_001"  // Deprecated.
     };
 }